// src/constants/types/searchUsers.types.ts
// Generated by https://quicktype.io

export interface IUsersSearch {
  birthCountry: BirthCountry
  firstName: string
  city: string
  favoriteRoles: string[]
  currentTeams: string[]
  lastName: string
  faceImage: string
  username: string
  type: string
  userId: string
  clubName: string
}

export interface BirthCountry {
  alpha2Code: string
  alpha3Code: string
  flag: string
  name: string
  phoneCode: string
  region: string
}